AF1_5 - Successful package installation support with pip
Now it is also possible to successfully install everything with pip (based on the example in the main README.md file). Slight changes were made to the pyproject.toml file.

In addition, slight changes were also made to the branch specific README.md file. As well as there should not be a need for .gitattributes file after all (at least for the time being).