AF2_7 - Branch specific README.md file
Filled the README.md file for this branch with relevant information based on the implemented changes. Additional things to note and comments were also added.

Along with this, some slight changes were made to the main manager classes - methods for creating new instances now have the arguments ordered in a more intuitive way.