GI [1.2.2] 1 - Preparation for update functionality testing
Slight changes to package version definition - it is now stored in separate VERSION.py file. Whenever any other part of this package requires the current version number, it will read it from this file.

In addition, this commit will allow testing of automatic updating feature with the command "interactive_noisy_simulation -u", which will then be added in a later commit.