GI [1.2.2] 3 - Message adjustments for terminal
Added slight styling functionality for output in terminal with the help of Python Rich. Now it is not just plain white text.

As a result of this, this package has gained 2x more dependencies:
- Python package "rich";
- Python package  "rich-argparse".

On top of this, all plain text that was previously stored in the python files is now located inside of messages.json.

Branch specific README.md file has been updated with information about main changes from this commit.