AF3_2 - Project directory restructuring
Because of the new changes, some restructuring had to be done to the project.

This commit moves all main functionality to a new folder called "core".

NOTE: Nothing in these files has changed, they have just been moved. This breaks all current INS functionality, however, this commit is being made so that it will be simpler to read through changes to each individual file further on.